Rename visualization and remove from v0
v0 visualization is redundant with visualization
