+++ src/spyglass/decoding/visualization/view1D.py
